backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
azure-identity==1.14.0
azure-mgmt-resource==23.0.1
